Decompose backend schema in preparation for custom auth implementation
