+++ backend/schema/auth.ts
@@ -1,0 +1,18 @@
+export type FilterArgs = {
+	session?: {
+		data: {id: string; isAdmin: boolean};
+	};
+} & Record<string, any>;
+
+export const isUserLoggedIn = (args: FilterArgs) => Boolean(args.session?.data.id);
+export const isUserAdmin = (args: FilterArgs) => Boolean(args.session?.data.isAdmin);
+
+export const isUserCurrentUserItem = (args: FilterArgs & {item: any}) =>
+	args.session?.data.id === args.item.id;
+export const isUserAdminOrCurrentUserItem = (args: FilterArgs & {item: any}) =>
+	isUserAdmin(args) || isUserCurrentUserItem(args);
+
+export const filterItemsOwnedByUser = (args: FilterArgs) => {
+	if (isUserAdmin(args)) return true;
+	return {owner: {id: {equals: args.session?.data.id}}};
+};
