+++ backend/schema/utils.ts
@@ -1,0 +1,4 @@
+export const requiredFieldConfig = {
+	validation: {isRequired: true},
+	graphql: {isNonNull: {read: true}},
+};
